refactored class component into functional, added components to separate files
