+++ src/pages/information/devteam/TeamMemberInfo.jsx
@@ -1,0 +1,82 @@
+const teamMembersData = [
+    {
+      id: 1,
+      name: "John Doe",
+      title: "CEO",
+      imgSrc:
+        "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQdEnuAJLWN_DdxE2nZyVUe2nfV7RknZMQXnw&s",
+      socialLinks: [
+        {
+          platform: "linkedin",
+          url: "https://www.linkedin.com/",
+        },
+      ],
+    },
+    {
+        id: 2,
+        name: "John Doe",
+        title: "CFO",
+        imgSrc:
+          "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQdEnuAJLWN_DdxE2nZyVUe2nfV7RknZMQXnw&s",
+        socialLinks: [
+          {
+            platform: "linkedin",
+            url: "https://www.linkedin.com/",
+          },
+        ],
+      },
+      {
+        id: 3,
+        name: "John Doe",
+        title: "Lead consultant",
+        imgSrc:
+          "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQdEnuAJLWN_DdxE2nZyVUe2nfV7RknZMQXnw&s",
+        socialLinks: [
+          {
+            platform: "linkedin",
+            url: "https://www.linkedin.com/",
+          },
+        ],
+      },
+      {
+        id: 4,
+        name: "John Doe",
+        title: "Senior Developer",
+        imgSrc:
+          "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQdEnuAJLWN_DdxE2nZyVUe2nfV7RknZMQXnw&s",
+        socialLinks: [
+          {
+            platform: "linkedin",
+            url: "https://www.linkedin.com/",
+          },
+        ],
+      },
+      {
+        id: 5,
+        name: "John Doe",
+        title: "Junior Developer",
+        imgSrc:
+          "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQdEnuAJLWN_DdxE2nZyVUe2nfV7RknZMQXnw&s",
+        socialLinks: [
+          {
+            platform: "linkedin",
+            url: "https://www.linkedin.com/",
+          },
+        ],
+      },
+      {
+        id: 6,
+        name: "John Doe",
+        title: "Junior Developer",
+        imgSrc:
+          "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQdEnuAJLWN_DdxE2nZyVUe2nfV7RknZMQXnw&s",
+        socialLinks: [
+          {
+            platform: "linkedin",
+            url: "https://www.linkedin.com/",
+          },
+        ],
+      }
+  ];
+
+export default teamMembersData;
